requirements.txt
SQLAlchemy
psycopg2
python-dotenv
argparse
astunparse
pandas
grizzly-sql
matplotlib